fix: alternativas to opcoes and new  js for Opções API
Troca da nomenclatura "Alternativa" para "opções", e criação do .js para os endpoints da API opções.
